lold base trait
