+++ src/solver/variable/CMakeLists.txt
@@ -75,6 +75,7 @@ source_group("variable\\adequacy" FILES ${SRC_VARIABLE_ADEQUACY})
 
 set(SRC_VARIABLE_ECONOMY
         include/antares/solver/variable/economy/all.h
+        include/antares/solver/variable/economy/economy_base.h
         include/antares/solver/variable/economy/links.h
 
         # Variables for Economy
@@ -116,8 +117,8 @@ set(SRC_VARIABLE_ECONOMY
         include/antares/solver/variable/economy/nbOfDispatchedUnits.h
         include/antares/solver/variable/economy/profitByPlant.h
         include/antares/solver/variable/economy/lold.h
-        include/antares/solver/variable/economy/loldCsr.h
         include/antares/solver/variable/economy/lold_base.h
+        include/antares/solver/variable/economy/loldCsr.h
         include/antares/solver/variable/economy/lolp.h
         include/antares/solver/variable/economy/nearPriceCap.h
         include/antares/solver/variable/economy/avail-dispatchable-generation.h
